Moved everything to lib to be able to expose internal interfaces and data
